Refactor PropertyBased tests to remove external dependencies
- Rename CompatibilityTests.cs to RandomPropertyBasedTests.cs for clarity
- Replace IntervalTree with LinearRangeFinder as reference implementation
- Move RangeDataGenerators.cs to Helper/ directory for better organization
- Update class documentation to emphasize random property-based testing
- Maintain RangeTree dependency only for dedicated compatibility tests
- All property tests now use internal LinearRangeFinder for validation
